+++ lib/pages/intro/intro_page.dart
@@ -2,8 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'package:flutter_linkify/flutter_linkify.dart';
 import 'package:go_router/go_router.dart';
-import 'package:url_launcher/url_launcher.dart';
-import 'package:url_launcher/url_launcher_string.dart';
+import 'package:extera_next/utils/url_launcher.dart';
 
 import 'package:extera_next/config/app_config.dart';
 import 'package:extera_next/generated/l10n/l10n.dart';
@@ -52,7 +51,7 @@ class IntroPage extends StatelessWidget {
                 ),
               ),
               PopupMenuItem(
-                onTap: () => launchUrl(Uri.parse(AppConfig.privacyUrl)),
+                onTap: () => openLink(AppConfig.privacyUrl),
                 child: Row(
                   mainAxisSize: .min,
                   children: [
@@ -109,7 +108,7 @@ class IntroPage extends StatelessWidget {
                           color: theme.colorScheme.secondary,
                           decorationColor: theme.colorScheme.secondary,
                         ),
-                        onOpen: (link) => launchUrlString(link.url),
+                        onOpen: (link) => openLink(link.url),
                       ),
                     ),
                     const Spacer(),