+++ lib/pages/settings/settings_view.dart
@@ -10,6 +10,7 @@ import 'package:extera_next/config/themes.dart';
 import 'package:extera_next/generated/l10n/l10n.dart';
 import 'package:extera_next/utils/fluffy_share.dart';
 import 'package:extera_next/utils/platform_infos.dart';
+import 'package:extera_next/utils/url_launcher.dart';
 import 'package:extera_next/widgets/avatar.dart';
 import 'package:extera_next/widgets/drawer.dart';
 import 'package:extera_next/widgets/list_divider.dart';
@@ -451,6 +452,25 @@ class SettingsView extends StatelessWidget {
                           const ListDivider(),
                           ListTile(
                             leading: CircleAvatar(
+                              backgroundColor: theme.colorScheme.primary,
+                              child: Icon(
+                                Icons.tune_outlined,
+                                color: theme.colorScheme.onPrimary,
+                              ),
+                            ),
+                            title: Text(L10n.of(context).customExperience),
+                            onTap: () =>
+                                context.go('/rooms/settings/custom_experience'),
+                            tileColor:
+                                activeRoute.startsWith(
+                                  '/rooms/settings/custom_experience',
+                                )
+                                ? theme.colorScheme.surfaceContainerHigh
+                                : null,
+                          ),
+                          const ListDivider(),
+                          ListTile(
+                            leading: CircleAvatar(
                               backgroundColor: theme.colorScheme.tertiary,
                               child: Icon(
                                 Icons.window_outlined,
@@ -581,7 +601,7 @@ class SettingsView extends StatelessWidget {
                               ),
                             ),
                             title: Text(L10n.of(context).privacy),
-                            onTap: () => launchUrlString(AppConfig.privacyUrl),
+                            onTap: () => openLink(AppConfig.privacyUrl),
                           ),
                           const ListDivider(),
                           ListTile(