+++ lib/utils/show_update_snackbar.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:url_launcher/url_launcher_string.dart';
+import 'package:extera_next/utils/url_launcher.dart';
 
 import 'package:extera_next/config/app_config.dart';
 import 'package:extera_next/generated/l10n/l10n.dart';
@@ -46,7 +46,7 @@ abstract class UpdateNotifier {
                       ),
                       onPressed: () {
                         ScaffoldMessenger.of(context).hideCurrentSnackBar();
-                        launchUrlString(AppConfig.donateUrl);
+                        openLink(AppConfig.donateUrl);
                       },
                       child: Text(L10n.of(context).supportDevelopment),
                     ),
@@ -59,7 +59,7 @@ abstract class UpdateNotifier {
                       ),
                       onPressed: () {
                         ScaffoldMessenger.of(context).hideCurrentSnackBar();
-                        launchUrlString(AppConfig.changelogUrl);
+                        openLink(AppConfig.changelogUrl);
                       },
                       icon: const Icon(Icons.list),
                       label: Text(L10n.of(context).changelog),